Update index.html
This code creates a form where the user can enter a task, with an "Add New Item" button that is disabled by default until the input field has text. Once the button is clicked, a new task is added to the list, with buttons to edit, save, delete, and mark it as done. The maximum number of tasks is defined as a constant MAX_TASKS, and the "Add New Item" button is disabled once this limit is reached. The "Clear All" button can be used to remove all tasks from the list.